==> 03_merge_english_and_french_to_estimate_word_and_concept_vocabulary_scores.Rmd <==
---
title: "03 Merge English and French to Estimate Word and Concept Vocabulary Scores"
author: "Miranda Gómez Díaz"
date: "2023-02-13"
output: html_document
---

```{r}
# Read needed packages
library(tidyverse)
```

```{r}
# Read files from previous step
Eng_02_df <- read_csv("data/02_Eng_keepers.csv")
Fr_02_df <- read_csv("data/02_Fr_keepers.csv")

# Read translation equivalents data
TE_list <- read_csv("data/ws_te_list_full.csv") %>%
  # remove unnecessary columns
  select(-c(word_en, word_fr, one_to_many)) %>%
  rename(eng_item_num = en_item_num)
```

```{r}
# Re estimate total_produced (sum all the produced words at a given month)
# Eng
Eng_df_total_produced <- Eng_02_df %>%
  # Create column for the sum the produced words
  mutate(eng_est_total_produced = rowSums(across(baa_baa:then), na.rm = TRUE)) %>% 
  # Rearrange column orders
  relocate(eng_est_total_produced, .after = age)
  
# Fr
Fr_df_total_produced <- Fr_02_df %>%
  # Create column for the sum the produced words
  mutate(fr_est_total_produced = rowSums(across(aie:si), na.rm = TRUE)) %>%
  # Rearrange column orders
  relocate(fr_est_total_produced, .after = age)
# S01, S06, and S81 have 0 in the first months because they did not know any words in French

```


```{r}
# Append item_id info to CDI data
# English
Eng_df_itemID <- Eng_df_total_produced %>%
  # Remove first 2 columns
  select(-c(sex: worried_boolean, eng_total_produced)) %>%
  # Convert to long data
  pivot_longer(-c(study_id:eng_est_total_produced), names_to = "english_item", values_to = "eng_acquired") %>%
  # Add item_id for each vocab
  group_by(study_id, n_months) %>%
  mutate(english_item_id = row_number(study_id)) %>%
  # Fix typo in english_item_id (*checked with eng_cdi from wordbank)
  mutate(english_item_id = case_when(english_item == "peas" ~ "131",
                                     english_item == "peanut_butter" ~ "132",
                                     english_item == "rocking_chair" ~ "282",
                                     english_item == "refrigerator" ~ "283",
                                     english_item == "window" ~ "292",
                                     english_item == "washing_machine" ~ "293",
                                     TRUE ~ as.character(english_item_id))) %>%
  mutate(english_item_id = as.numeric(english_item_id)) %>%
  ungroup() %>%
  select(study_id, TRUE_months, n_months, completed, eng_est_total_produced, age, english_item, english_item_id, eng_acquired) %>% 
  mutate(item_id = english_item_id) %>% # LAIA: adding this line here for when we need to merge TEs
  # Rename columns for when they are joined with French
  rename(eng_item_num = english_item_id,
         age_eng = age,
         eng_completed = completed) %>%
  # Append translation equivalents (TE) list to data
  left_join(TE_list, by = "eng_item_num")

# French
Fr_df_itemID <- Fr_df_total_produced %>%
  # Remove first 2 columns
  select(-c(sex: worried_boolean, fr_total_produced)) %>%
  # Convert to long data
  pivot_longer(-c(study_id:fr_est_total_produced), names_to = "french_item", values_to = "fr_acquired") %>%
  # Add item_id for each vocab
  group_by(study_id, n_months) %>%
  mutate(french_item_id = row_number(study_id)) %>%
  mutate(french_item_id = as.numeric(french_item_id)) %>%
  ungroup() %>%
  select(study_id, TRUE_months, n_months, completed, fr_est_total_produced, age, french_item, french_item_id, fr_acquired) %>% 
  mutate(item_id = french_item_id) %>% # LAIA: adding this line here for when we need to merge TEs
  # Rename columns for when they are joined with English
  rename(fr_item_num = french_item_id,
         age_fr = age,
         fr_completed = completed) %>%
  # Append translation equivalents (TE) list to data
  left_join(TE_list, by = "fr_item_num")

```

```{r}
# Merge Eng and Fr dataframes 

EngFr_df <- full_join(Eng_df_itemID, Fr_df_itemID, by = c("study_id", "n_months", "te_id", "eng_item_num", "fr_item_num"), suffix = c("_eng", "_fr")) %>%
  # Replace with 0 the NAs in total produced in eng and fre because those children are in fact monolinguals.
  mutate(eng_est_total_produced = case_when(is.na(eng_est_total_produced) ~ 0,
                           TRUE ~ eng_est_total_produced),
         fr_est_total_produced = case_when(is.na(fr_est_total_produced) ~ 0,
                           TRUE ~ fr_est_total_produced)) %>%
  # Replace with NA the 0s in est_total_produced when a bilingual baby did not complete the CDI in one language, but did in the other
  mutate(eng_est_total_produced = case_when(eng_completed == FALSE ~ as.numeric(NA), 
                                            TRUE ~ eng_est_total_produced),
         fr_est_total_produced = case_when(fr_completed == FALSE ~ as.numeric(NA), 
                                            TRUE ~ fr_est_total_produced),
         age_eng = case_when(age_eng == 999 ~ as.numeric(NA),
                             TRUE ~ age_eng),
         age_fr = case_when(age_fr == 999 ~ as.numeric(NA),
                             TRUE ~ age_fr),
         num_datapts = case_when(is.na(TRUE_months_eng) ~ TRUE_months_fr,
                                 is.na(TRUE_months_fr) ~ TRUE_months_eng,
                                 TRUE_months_eng > TRUE_months_fr ~ TRUE_months_eng,
                                 TRUE_months_fr > TRUE_months_eng ~ TRUE_months_fr,
                                 TRUE ~ TRUE_months_eng))

# Fix age (Due to the nature of the administration process where the questionnaires are sent on the 1st of each month, there would be overlaps in the age when the questionnaires are filled. So when generating the report, we rely on the n_month variable to generate an estimate age to avoid overlaps.)
EngFr_df_agefixed <- EngFr_df %>% 
  # Sometimes parents may have missed a language (i.e., no data), so replace NA in a missing language with the age in the other language
     mutate(age_fixed = case_when(is.na(age_eng) ~ age_fr, 
                           is.na(age_fr) ~ age_eng,
                            # When both languages are available, compare the age across two languages and keep the smaller number (i.e., the age they first joined the project)
                           age_eng < age_fr ~ age_eng,
                           age_eng > age_fr ~ age_fr,
                           # Otherwise, return the age in the English form (since this should be the same as the age in the French form)
                           TRUE ~ age_eng)) %>% 
  group_by(study_id) %>%
  mutate(starting_age = min(age_fixed, na.rm = TRUE)) %>% # Extract the minimum age as the starting age of the project
  ungroup() %>%
  mutate(age = age_fixed) 

```

```{r}
# Estimate word and concept vocabulary scores
# Word vocabulary score
word_concept_scores <- EngFr_df_agefixed %>%
  mutate(words_produced = rowSums(across(c(eng_est_total_produced, fr_est_total_produced)), na.rm = FALSE)) %>%
 # Rearrange column orders
  relocate(words_produced, .after = n_months) %>%
  relocate(eng_est_total_produced, .after = words_produced) %>%
  relocate(fr_est_total_produced, .after = eng_est_total_produced) %>%
  # Estimate number of TEs produced
  mutate(te_acquired = case_when(eng_acquired == 1 & fr_acquired == 1 ~ 1,
                                 TRUE ~ 0)) %>%
  group_by(study_id, n_months) %>%
  mutate(te_produced = sum(te_acquired)) %>%
  ungroup() %>%
  # Concept vocabulary score
  mutate(concepts_produced = words_produced - te_produced) %>%
  relocate(concepts_produced, .after = words_produced) %>%
  relocate(te_produced, .after = concepts_produced) %>% 
  # Remove unnecessary columns
  relocate(fr_completed, .after = eng_completed) %>%
  select(study_id, n_months:fr_completed, num_datapts, age, starting_age, item_id_eng) 

```

```{r}
# Save dataframe as csv file
write_csv(word_concept_scores, 'data/03_vocabulary_scores.csv')
```
