==> 05.1_identified_and_unidentified_vocabulary_spurt.Rmd <==
---
title: "05.1 Identifiable and Unidentifiable Vocabulary Spurt"
author: "Miranda Gómez Díaz"
date: "2023-05-18"
output: html_document
---

```{r}
#Read needed packages
library(tidyverse)
library(ez)
library(report)
```

```{r}
# Read files from previous steps
coefficients_04_df <- read_csv("data/04_analysis_step_1_coefficients.csv")
LEQ_01_df <- read_csv("data/01_LEQ_columns.csv")
vocabulary_scores_df <- read_csv("data/04_analysis_step_1_dataset.csv") 
final_exposure_df <- read_csv("data/05_final_exposure_dataset.csv")
final_vocabulary_scores_df <- read_csv("data/05_final_vocabulary_scores_dataset.csv")
words_at_spurt_distinct <- read_csv('data/05_words_at_spurt_distinct_dataset.csv')

# These babies had been excluded in one but not the other language (they had missing data or exposure change in only one language)
to_be_excluded <- read_csv("data/02_total_excluded.csv")
```

```{r}
all_scores_df <- vocabulary_scores_df %>% # Some babies do not have data in one language (English or French), it is either because they are monolingual or because their data in that language was removed due to missing data or exposure change in that language
  # Estimate number of data points for each baby
  group_by(study_id, vocabulary_type_step1) %>% 
  mutate(completed_months = sum(!(is.na(vocabulary_score)))) %>%
  # Join with coefficients dataframe
  full_join(coefficients_04_df, by = c("study_id", "vocabulary_type_step1")) %>% 
  ungroup() %>% 
  # Join coefficients and language exposure dataframes
  # Use left_join because LEQ dataset has full sample (N=125), as babies were not excluded from it
  group_by(study_id) %>% 
  left_join(LEQ_01_df, by = "study_id") %>% 
  ungroup() %>% 
  # Remove word and concept data from babies who were excluded in one language due to missing data or exposure change in that language
  filter(!((vocabulary_type_step1 == "word" | vocabulary_type_step1 == "concept") & study_id %in% to_be_excluded$study_id)) %>% 
  # Set L3 and L4 exposure as 0 if babies had no exposure (NA)
  mutate(cumulative_exp_L3 = case_when(is.na(cumulative_exp_L3) ~ 0,
                                       TRUE ~ cumulative_exp_L3),
         cumulative_exp_L4 = case_when(is.na(cumulative_exp_L4) ~ 0,
                                       TRUE ~ cumulative_exp_L4)) %>% 
   mutate(vocabulary_type_step1 = factor(vocabulary_type_step1, #HK: removed the across() here, for same reason as in script 05 line 63
                levels=c("English", "French", "concept", "word"))) 
  
```

```{r}
# Determine if the midpoint is identified (within the age range of the study) or not
midpoint_identification <- all_scores_df %>% 
  group_by(study_id) %>% 
  mutate(midp_identify = case_when(study_id == "S125" & vocabulary_type_step1 == "concept" ~ "not_identified", # Manually set the baby as non-spurter (very small spurt)
                                   midpoint>=starting_age & midpoint<=max(age) ~ "identified",
                                   midpoint<starting_age | midpoint>max(age) ~ "not_identified"))

# Keep one row per baby per vocabulary type
midpoint_identification_sum <- midpoint_identification %>% 
  group_by(study_id, vocabulary_type_step1) %>% 
  filter(vocabulary_score == max(vocabulary_score, na.rm = TRUE)) %>% 
  slice_head() %>% #HK note: this section was creating a dataframe where S118 and S033 had 2 rows each for concept vocabulary, I think because both rows had the same max vocab that you were filtering to on the previous row. This line will keep just one row, as stated in your comment in line 61.
  ungroup() %>% #HK note: adding ungroup() here 
  mutate(midp_identify =as.factor(midp_identify))
```

```{r}
# Estimate vocabulary size by age with logistic equation

#Join coefficients and full vocabulary data
logistic_df <- coefficients_04_df %>%
   # Join vocabulary scores dataframe
  full_join(vocabulary_scores_df, by = c("study_id", "vocabulary_type_step1")) %>%
  group_by(study_id) %>%
  # Remove word and concept data from babies who were excluded in one language due to missing data or exposure change in that language
  filter(!((vocabulary_type_step1 == "word" | vocabulary_type_step1 == "concept") & study_id %in% to_be_excluded$study_id)) %>%
  mutate(id_voc = paste0(study_id, vocabulary_type_step1)) %>% # This will be used to filter out data in another df that was removed here
  # Round midpoint to one decimal place
  mutate(midpoint = round(midpoint, digits = 2))

# Create dataframe with age for the logistic curve
ages <- tibble(age = seq(16, 30, by = 0.01))
id_voc <- logistic_df %>% distinct(study_id, vocabulary_type_step1, id_voc)
id_voc_ages <- merge(id_voc, ages)

# Create the logistic curve with the equation's parameters
curve <- id_voc_ages %>%
  left_join(coefficients_04_df, by = c("study_id", "vocabulary_type_step1")) %>%
  mutate(vocabulary_type_step1 = factor(vocabulary_type_step1,
                levels=c("English", "French", "word", "concept"))) %>%
  mutate(vocabulary_score = upper_limit / (1 + exp(-slope*(log(age) - log(midpoint)))),
         centered_age = age - midpoint,
         midpoint = round(midpoint, digits = 2)) %>%
  # Add midpoint identification column
  left_join(midpoint_identification_sum %>% select(study_id, vocabulary_type_step1, midp_identify), by = c("study_id", "vocabulary_type_step1"))
```

```{r}
# Create df with vocabulary size at 18 months
vocab_at_18m <- 
  # curve %>% filter(age == 18) # Use this line to use vocabulary size estimated by the logistic equation
  midpoint_identification %>% filter(age >= 18 & age < 19) %>% # Use this line to use actual vocabulary size (there might be babies with missing data)
  group_by(study_id, vocabulary_type_step1) %>%
  filter(row_number() == 1) %>% 
  # Remove columns
  ungroup()  

# Vocabulary size at 18 months descriptives
vocab_at_18m_summary <- vocab_at_18m %>%
  group_by(study_id, vocabulary_type_step1) %>%
  filter(row_number() == 1) %>% 
  # Remove columns
  ungroup() %>% 
    # Summary
  group_by(vocabulary_type_step1, midp_identify) %>%
  summarise(mean_score = mean(vocabulary_score, na.rm = T), #HK note: a lot of these are coming up NA for the unidentified ones, so I added in na.rm = TRUE to base the summary stats on the available data, and I did this for all lines up to 118
            SD_score = sd(vocabulary_score, na.rm = T),
            min_score = min(vocabulary_score, na.rm = T),
            max_score = max(vocabulary_score, na.rm = T)) %>% 
  # Round  to two decimal place
  mutate(mean_score = round(mean_score, digits = 2),
         SD_score = round(SD_score, digits = 2),
         min_score = round(min_score, digits = 2),
         max_score = round(max_score, digits = 2))

# Create df with vocabulary size at 30 months
vocab_at_30m <- 
  # curve %>% filter(age == 30) # Use this line to use vocabulary size estimated by the logistic equation
  # Use these lines to use actual vocabulary size (there might be babies with missing data)
  midpoint_identification %>%
  group_by(study_id, vocabulary_type_step1) %>%
  arrange(desc(age)) %>%
  filter(row_number() == 1)

# Vocabulary size at 30 months descriptives
vocab_at_30m_summary <- vocab_at_30m %>%
  group_by(study_id, vocabulary_type_step1) %>%
  filter(row_number() == 1) %>% 
  # Remove columns
  ungroup() %>% 
  # Summary
  group_by(vocabulary_type_step1, midp_identify) %>% 
  summarise(mean_score = mean(vocabulary_score, na.rm = T), #HK note: I added na.rm = TRUE for lines 142-145
            SD_score = sd(vocabulary_score, na.rm = T),
            min_score = min(vocabulary_score, na.rm = T),
            max_score = max(vocabulary_score, na.rm = T)) %>% 
  # Round  to two decimal place
  mutate(mean_score = round(mean_score, digits = 2),
         SD_score = round(SD_score, digits = 2),
         min_score = round(min_score, digits = 2),
         max_score = round(max_score, digits = 2))
```

```{r}
# t-tests for English vocabulary
# Vocabulary size at 18 m
Eng_18m_df <- vocab_at_18m %>% 
  filter(vocabulary_type_step1 == "English")

vocab_english_18m_test <- t.test(vocabulary_score ~ midp_identify, data = Eng_18m_df)
report(vocab_english_18m_test, digits = 3)

# Vocabulary size at 30 m
Eng_30m_df <- vocab_at_30m %>% 
  filter(vocabulary_type_step1 == "English")

vocab_english_30m_test <- t.test(vocabulary_score ~ midp_identify, data = Eng_30m_df)
report(vocab_english_30m_test, digits = 3)

```


```{r}
# t-tests for French vocabulary
# Vocabulary size at 18 m
Fr_18m_df <- vocab_at_18m %>% 
  filter(vocabulary_type_step1 == "French")

vocab_french_18m_test <- t.test(vocabulary_score ~ midp_identify, data = Fr_18m_df)
report(vocab_french_18m_test, digits = 3)

# Vocabulary size at 30 m
Fr_30m_df <- vocab_at_30m %>% 
  filter(vocabulary_type_step1 == "French")

vocab_french_30m_test <- t.test(vocabulary_score ~ midp_identify, data = Fr_30m_df)
report(vocab_french_30m_test, digits = 3)

```

```{r}
# t-tests for word vocabulary
# Vocabulary size at 18 m
word_18m_df <- vocab_at_18m %>% 
  filter(vocabulary_type_step1 == "word")

vocab_word_18m_test <- t.test(vocabulary_score ~ midp_identify, data = word_18m_df)
report(vocab_word_18m_test, digits = 3)

# Vocabulary size at 30 m
word_30m_df <- vocab_at_30m %>% 
  filter(vocabulary_type_step1 == "word")

vocab_word_30m_test <- t.test(vocabulary_score ~ midp_identify, data = word_30m_df)
report(vocab_word_30m_test, digits = 3)

```

```{r}
# t-tests for concept vocabulary
# Vocabulary size at 18 m
concept_18m_df <- vocab_at_18m %>% 
  filter(vocabulary_type_step1 == "concept")

vocab_concept_18m_test <- t.test(vocabulary_score ~ midp_identify, data = concept_18m_df)
report(vocab_concept_18m_test, digits = 3)

# Vocabulary size at 30 m
concept_30m_df <- vocab_at_30m %>% 
  filter(vocabulary_type_step1 == "concept")

vocab_concept_30m_test <- t.test(vocabulary_score ~ midp_identify, data = concept_30m_df)
report(vocab_concept_30m_test, digits = 3)

```

```{r}
# Patterns of bilingual spurt (recreate step2_df, but including non-spurters)
spurt_patterns_by_dominance <- coefficients_04_df %>%
  # Join coefficients and language exposure dataframes
  # Use left_join because LEQ dataset has full sample (N=125), as babies were not excluded from it
  left_join(final_exposure_df, by = "study_id") %>%
  mutate(vocabulary_type_step2 = case_when(
    vocabulary_type_step1 == "English" | vocabulary_type_step1 == "French" ~ "single",
    vocabulary_type_step1 == "word" ~ "word",
    vocabulary_type_step1 == "concept" ~ "concept")) %>%
  relocate(vocabulary_type_step2, .after = vocabulary_type_step1) %>%
  mutate(exposure_percent = case_when(
    vocabulary_type_step1 == "English" ~ cumulative_exp_Eng,
    vocabulary_type_step1 == "French" ~ cumulative_exp_Fre)) %>% 
  # Join vocabulary scores dataframe
  left_join(final_vocabulary_scores_df, by = c("study_id", "vocabulary_type_step1")) %>% 
  # # Remove babies with midpoint out of age range
  # filter(midpoint>=16 & midpoint<=age) %>%
  # Set language dominance
  mutate(dominance = case_when(exposure_percent>50 ~ "dominant",
                               exposure_percent<50 ~ "nondominant")) %>% 
  # Remove word and concept data from babies who were excluded in one language due to missing data or exposure change in that language
  filter(!((vocabulary_type_step1 == "word" | vocabulary_type_step1 == "concept") & study_id %in% to_be_excluded$study_id)) %>% 
  # Join with words learned at spurt dataframe
  left_join(words_at_spurt_distinct, by = c("study_id", "vocabulary_type_step1")) %>%
  group_by(study_id) %>% 
  mutate(midp_identify = case_when(midpoint>=starting_age & midpoint<=max(age) ~ "identified",
                                   midpoint<starting_age | midpoint>max(age) ~ "not_identified")) %>% 
  # Keep only single-language data
  filter(vocabulary_type_step2 == "single") %>% 
  # Categorize by lang_status and spurter/non-spurter
  mutate(lang_spurt_status = case_when((lang_status == "fr_mono" | lang_status == "eng_mono") & midp_identify == "identified" ~ "mono_spurter",
                                       (lang_status == "fr_mono" | lang_status == "eng_mono") & midp_identify == "not_identified" ~ "mono_nonspurter",
                                       lang_status == "bilingual" & dominance == "dominant" & midp_identify == "identified" ~ "biling_dominant_spurter",
                                       lang_status == "bilingual" & dominance == "dominant" & midp_identify == "not_identified" ~ "biling_dominant_nonspurter",
                                       lang_status == "bilingual" & dominance == "nondominant" & midp_identify == "identified" ~ "biling_nondominant_spurter",
                                       lang_status == "bilingual" & dominance == "nondominant" & midp_identify == "not_identified" ~ "biling_nondominant_nonspurter")) %>% 
  select(study_id, exposure_percent, lang_status, words_learned, dominance, lang_spurt_status, midpoint, slope) %>% 
  ungroup() #HK note: added ungroup here to reset the data frame

# Descriptives of timing and words learned at spurt by language status/dominance group
spurt_patterns_descriptives <- spurt_patterns_by_dominance %>% 
  # # Remove non-spurters
  # filter(lang_spurt_status == "mono_spurter" |
  #       lang_spurt_status == "biling_dominant_spurter" |
  #       lang_spurt_status == "biling_nondominant_spurter") %>% 
  mutate(lang_status = str_replace_all(lang_status, "fr_mono", "monolingual"),
         lang_status = str_replace_all(lang_status, "eng_mono", "monolingual")) %>% 
  group_by(lang_status, lang_spurt_status, dominance) %>% 
  summarise(lang_spurt.N = n(),
            midpoint.M = round(mean(midpoint), digits = 2),
            midpoint.SD = round(sd(midpoint), digits = 2),
            slope.M = round(mean(slope), digits = 2),
            slope.SD = round(sd(slope), digits = 2),
            exposure.M = round(mean(exposure_percent), digits = 2),
            exposure.SD = round(sd(exposure_percent), digits = 2),
            wordslearned.M = round(mean(words_learned), digits = 2),
            wordslearned.SD = round(sd(words_learned), digits = 2)
            ) %>% 
  ungroup() %>% 
  group_by(lang_status, dominance) %>% 
  mutate(lang_dom.N = sum(lang_spurt.N), # Estimate number of babies in each language status by dominance group (e.g., bilingual-dominant, bilingual-nondominant)
         lang_spurt.percent = round((lang_spurt.N * 100 / lang_dom.N), digits = 0)) %>%  # Estimate percentage of spurters and non-spurters in each anguage status by dominance group
  relocate(lang_dom.N, .after = dominance) %>% 
  relocate(lang_spurt.percent, .after = lang_spurt.N)

```
```{r}
# Bilinguals' spurt patterns
bilingual_spurt_pattern <- spurt_patterns_by_dominance %>% 
  filter(lang_status == "bilingual") %>% 
  select(study_id, dominance, lang_spurt_status) %>% 
  # separate(lang_spurt_status, c(NA, "dominance", "spurt_status"), sep = "_") %>% 
  pivot_wider(id_cols = study_id, names_from = dominance, values_from = lang_spurt_status) %>% 
  # Remove babies with data in only one language
  filter(!is.na(nondominant)) %>% 
  # Determine in how many languages a baby spurted
  mutate(num_spurts = case_when(dominant == "biling_dominant_spurter" & nondominant == "biling_nondominant_spurter" ~ "both",
                                dominant == "biling_dominant_spurter" & nondominant == "biling_nondominant_nonspurter" ~ "one",
                                dominant == "biling_dominant_nonspurter" & nondominant == "biling_nondominant_spurter" ~ "one",
                                dominant == "biling_dominant_nonspurter" & nondominant == "biling_nondominant_nonspurter" ~ "none"))
```

```{r}
# Create dfs for bilingual and monolingual spurters

# Bilingual spurters
bilingual_spurters_df <- spurt_patterns_by_dominance %>% 
  filter(lang_spurt_status == "biling_dominant_spurter" | lang_spurt_status == "biling_nondominant_spurter") %>%
  select(study_id, midpoint, dominance, words_learned)
  
# Estimate difference in timing of and words learned during the spurt
bilingual_spurters_by_dominance_df <- bilingual_spurters_df%>% 
  pivot_wider(names_from = dominance, values_from = c(midpoint, words_learned)) %>% 
  mutate(midpoint_diff = midpoint_dominant - midpoint_nondominant, 
         words_learned_diff = words_learned_dominant - words_learned_nondominant) %>% 
  ungroup() %>% 
  summarise(midpoint_diff.M = round(mean(midpoint_diff, na.rm = TRUE), digits = 2),
            midpoint_diff.SD = round(sd(midpoint_diff, na.rm = TRUE), digits = 2),
            words_learned_diff.M = round(mean(words_learned_diff, na.rm = TRUE), digits = 0),
            words_learned_diff.SD = round(sd(words_learned_diff, na.rm = TRUE), digits = 0))

# Monolingual spurters
monolingual_spurters_df <- spurt_patterns_by_dominance %>% 
  filter(lang_spurt_status == "mono_spurter") %>%
  select(study_id, midpoint, words_learned) 
```

```{r}
# Descriptives of bilingual and monolingual spurters (midpoint and words learned during spurt)

# Bilingual spurters
bilingual_spurters_descriptives <- bilingual_spurters_df %>% 
  group_by(dominance) %>% 
  summarise(midpoint.M = round(mean(midpoint), digits = 2),
            midpoint.SD = round(sd(midpoint), digits = 2),
            words_learned.M = round(mean(words_learned), digits = 0),
            words_learned.SD = round(sd(words_learned), digits = 0))

# Monolingual spurters
monolingual_spurters_descriptives <- monolingual_spurters_df %>% 
  ungroup() %>% 
  summarise(midpoint.M = round(mean(midpoint), digits = 2),
            midpoint.SD = round(sd(midpoint), digits = 2),
            words_learned.M = round(mean(words_learned), digits = 0),
            words_learned.SD = round(sd(words_learned), digits = 0))

```

```{r}
# Compare language exposure bilingual spurters vs nonspurters in their dominant language
bilingual_dominant_df <- spurt_patterns_by_dominance %>% 
  filter(lang_status == "bilingual",
         dominance == "dominant") %>% 
  mutate(lang_spurt_status = factor(lang_spurt_status,
                levels=c("biling_dominant_spurter", "biling_dominant_nonspurter")))

bilingual_dominant_test <- t.test(exposure_percent ~ lang_spurt_status, data = bilingual_dominant_df)
report(bilingual_dominant_test, digits = 3)

```

```{r}
# Compare language exposure bilingual spurters vs nonspurters in their nondominant language
bilingual_nondominant_df <- spurt_patterns_by_dominance %>% 
  filter(lang_status == "bilingual",
         dominance == "nondominant") %>% 
  mutate(lang_spurt_status = factor(lang_spurt_status,
                levels=c("biling_nondominant_spurter", "biling_nondominant_nonspurter")))

bilingual_nondominant_test <- t.test(exposure_percent ~ lang_spurt_status, data = bilingual_nondominant_df)
report(bilingual_nondominant_test, digits = 3)

```

```{r}
# Save dataframe as csv file
write_csv(midpoint_identification, 'data/05.1_midpoint_identification.csv')
write_csv(midpoint_identification_sum, 'data/05.1_midpoint_identification_summary.csv')
```
