==> 05.3_control_demographic_variables.Rmd <==
---
title: "05.3_control_demographic_variables"
author: "Miranda Gómez Díaz"
date: "`r Sys.Date()`"
output: html_document
---

```{r}
#Read needed packages
library(tidyverse)
library(tidylog)
library(rempsyc)
library(lmerTest)
library(lme4)
library(performance)
library(report)
library(jtools)
library(sjPlot)
library(gridExtra)
library(cowplot)
library(ggpubr)
library(janitor)
```

```{r}
# Read csv files
step2_df <- read_csv('data/05_analysis_step_2_dataset.csv')
demographics_df <- read_csv('data/07_demographics_dataset.csv')
```

```{r}
# Filter demographics for infants in step2_df
step2_demographics_df <- demographics_df %>% 
  filter(study_id %in% step2_df$study_id) %>% 
  # Convert income ranges into the range average. "Prefer not to disclose" will be considered NA.
  mutate(income_range_average = case_when(correct_income == "25000-50000" ~ (25000+50000)/2,
                                     correct_income == "50000-75000" ~ (50000+75000)/2,
                                     correct_income == "75000-100000" ~ (75000+100000)/2,
                                     correct_income == "100000-125000" ~ (100000+125000)/2,
                                     correct_income == "125000-150000" ~ (125000+150000)/2,
                                     correct_income == "150000-175000" ~ (150000+175000)/2,
                                     correct_income == "175000-200000" ~ (175000+200000)/2,
                                     correct_income == ">200000" ~ (200000+225000)/2),
    correct_birth_order = as.factor(correct_birth_order),
    birth_order_categorical = case_when(correct_birth_order == "1" ~ "first_born",
                                        correct_birth_order == "0" ~ NA,
                                        TRUE ~ "later_born"),
    maternal_education_z = correct_mother_education - mean(correct_mother_education) / sd(correct_mother_education),
    income_z = income_range_average - mean(income_range_average, na.rm = TRUE) / sd(income_range_average, na.rm = TRUE),
    sex_s = case_when(correct_sex == "Female" ~ -0.5,
                      correct_sex == "Male" ~ 0.5),
    birth_order_s = case_when(birth_order_categorical == "first_born" ~ -0.5,
                              birth_order_categorical == "later_born" ~ 0.5)
    # sex_s = c(scale(correct_sex, center = TRUE, scale = TRUE)),
    # birth_order_s = c(scale(birth_order_categorical, center = TRUE, scale = TRUE))
         # # Estimate parental education
         # parental_education = case_when(is.na(correct_father_education) ~ correct_mother_education,
         #                                TRUE ~ (correct_mother_education + correct_father_education) / 2
         # # parental_education = case_when(is.na(correct_father_education) ~ correct_mother_education)
         )
```

```{r}
# Visualize SES and parental education 

av_income_hist <- hist(step2_demographics_df$income_range_average)

mother_education_hist <- hist(step2_demographics_df$correct_mother_education)

father_education_hist <- hist(step2_demographics_df$correct_father_education)
```

```{r}
# Correlations between income and parental education

ggscatter(step2_demographics_df, x = "income_range_average", y = "correct_mother_education", 
          add = "reg.line", 
          conf.int = TRUE, 
          cor.coef = TRUE, 
          cor.method = "pearson",
          xlab = "Annual income average range", ylab = "Mother education")

ggscatter(step2_demographics_df, x = "income_range_average", y = "correct_father_education", 
          add = "reg.line", 
          conf.int = TRUE, 
          cor.coef = TRUE, 
          cor.method = "pearson",
          xlab = "Annual income factor", ylab = "Father education")

ggscatter(step2_demographics_df, x = "income_range_average", y = "parental_education", 
          add = "reg.line", 
          conf.int = TRUE, 
          cor.coef = TRUE, 
          cor.method = "pearson",
          xlab = "Annual income factor", ylab = "Parental education")

ggscatter(step2_demographics_df, x = "correct_mother_education", y = "correct_father_education", 
          add = "reg.line", 
          conf.int = TRUE, 
          cor.coef = TRUE, 
          cor.method = "pearson",
          xlab = "Mother education", ylab = "Father education")

```

```{r}
# Join dataframes
step2_full_df <- step2_df %>% 
  left_join(step2_demographics_df, by = "study_id")

write_csv(step2_full_df, 'data/05.3_demographics_analysis_step_2_dataset.csv')

# Prepare dataframes for models

# Single-language vocabulary
single_lang_df <- step2_full_df %>%
  filter(vocabulary_type_step2 == "single") 
# Concept vocabulary
concept_df <- step2_full_df %>%
  filter(vocabulary_type_step2 == "concept")
# Word vocaublary
word_df <- step2_full_df %>%
  filter(vocabulary_type_step2 == "word")
```

```{r}
# Single-language midpoint model

# All demographic variables
midpoint_single_model <- 
  lmer(midpoint ~ exposure_percent + 
         correct_sex +
         income_range_average  + 
         correct_mother_education + 
         # correct_father_education + 
         birth_order_categorical +
         (1|study_id), data = single_lang_df)

check_model(midpoint_single_model)
model_performance(midpoint_single_model)
summary(midpoint_single_model)
summ(midpoint_single_model, digits = 3)
report(midpoint_single_model, digits = 3)

# Only sex
midpoint_single_model_sex <- 
  lmer(midpoint ~ exposure_percent + 
         correct_sex +
         (1|study_id), data = single_lang_df)

check_model(midpoint_single_model_sex)
model_performance(midpoint_single_model_sex)
summary(midpoint_single_model_sex)
summ(midpoint_single_model_sex, digits = 3)
report(midpoint_single_model_sex, digits = 3)

# Only income
midpoint_single_model_income <- 
  lmer(midpoint ~ exposure_percent + 
         income_range_average +
         (1|study_id), data = single_lang_df)

check_model(midpoint_single_model_income)
model_performance(midpoint_single_model_income)
summary(midpoint_single_model_income)
summ(midpoint_single_model_income, digits = 3)
report(midpoint_single_model_income, digits = 3)

# Only mother education
midpoint_single_model_mother <- 
  lmer(midpoint ~ exposure_percent * 
         correct_mother_education +
         (1|study_id), data = single_lang_df)

check_model(midpoint_single_model_mother)
model_performance(midpoint_single_model_mother)
summary(midpoint_single_model_mother)
summ(midpoint_single_model_mother, digits = 3)
report(midpoint_single_model_mother, digits = 3)

# # Only father education
# midpoint_single_model_father <- 
#   lmer(midpoint ~ exposure_percent + 
#          correct_father_education +
#          (1|study_id), data = single_lang_df)
# 
# check_model(midpoint_single_model_father)
# model_performance(midpoint_single_model_father)
# summary(midpoint_single_model_father)
# summ(midpoint_single_model_father, digits = 3)
# report(midpoint_single_model_father, digits = 3)

# Only birth order
midpoint_single_model_birth_order <- 
  lmer(midpoint ~ exposure_percent + 
         birth_order_categorical +
         (1|study_id), data = single_lang_df)

check_model(midpoint_single_model_birth_order)
model_performance(midpoint_single_model_birth_order)
summary(midpoint_single_model_birth_order)
summ(midpoint_single_model_birth_order, digits = 3)
report(midpoint_single_model_birth_order, digits = 3)
```

```{r}
# Single-language slope model

# All demographic variables
slope_single_model <- 
  lmer(slope ~ exposure_percent + 
         correct_sex +
         income_range_average +
         correct_mother_education +
         # correct_father_education + 
         birth_order_categorical +
         (1|study_id), data = single_lang_df)

check_model(slope_single_model)
model_performance(slope_single_model)
summary(slope_single_model)
summ(slope_single_model, digits = 3)
report(slope_single_model, digits = 3)

# Only sex
slope_single_model_sex <- 
  lmer(slope ~ exposure_percent + 
         correct_sex +
         (1|study_id), data = single_lang_df)

check_model(slope_single_model_sex)
model_performance(slope_single_model_sex)
summary(slope_single_model_sex)
summ(slope_single_model_sex, digits = 3)
report(slope_single_model_sex, digits = 3)

# Only income
slope_single_model_income <- 
  lmer(slope ~ exposure_percent + 
         income_range_average +
         (1|study_id), data = single_lang_df)

check_model(slope_single_model_income)
model_performance(slope_single_model_income)
summary(slope_single_model_income)
summ(slope_single_model_income, digits = 3)
report(slope_single_model_income, digits = 3)

# Only mother education
slope_single_model_mother <- 
  lmer(slope ~ exposure_percent + 
         correct_mother_education +
         (1|study_id), data = single_lang_df)

check_model(slope_single_model_mother)
model_performance(slope_single_model_mother)
summary(slope_single_model_mother)
summ(slope_single_model_mother, digits = 3)
report(slope_single_model_mother, digits = 3)

# # Only father education
# slope_single_model_father <- 
#   lmer(slope ~ exposure_percent + 
#          correct_father_education +
#          (1|study_id), data = single_lang_df)
# 
# check_model(slope_single_model_father)
# model_performance(slope_single_model_father)
# summary(slope_single_model_father)
# summ(slope_single_model_father, digits = 3)
# report(slope_single_model_father, digits = 3)

# Only birth order
slope_single_model_birth_order <- 
  lmer(slope ~ exposure_percent + 
         birth_order_categorical +
         (1|study_id), data = single_lang_df)

check_model(slope_single_model_birth_order)
model_performance(slope_single_model_birth_order)
summary(slope_single_model_birth_order)
summ(slope_single_model_birth_order, digits = 3)
report(slope_single_model_birth_order, digits = 3)
```

```{r}
# Concept vocabulary midpoint model

# All demographic variables
midpoint_concept_model <- 
  lm(midpoint ~ exposure_balance +
       correct_sex +
       income_range_average +
       correct_mother_education +
       # correct_father_education + 
       birth_order_categorical,
     data = concept_df)

check_model(midpoint_concept_model)
model_performance(midpoint_concept_model)
summary(midpoint_concept_model)
summ(midpoint_concept_model, digits = 3)
report(midpoint_concept_model, digits = 3)

# Only sex
midpoint_concept_model_sex <- 
  lm(midpoint ~ exposure_balance * 
         correct_sex,
       data = concept_df)

check_model(midpoint_concept_model_sex)
model_performance(midpoint_concept_model_sex)
summary(midpoint_concept_model_sex)
summ(midpoint_concept_model_sex, digits = 3)
report(midpoint_concept_model_sex, digits = 3)

# Only income
midpoint_concept_model_income <- 
  lm(midpoint ~ exposure_balance + 
         income_range_average,
       data = concept_df)

check_model(midpoint_concept_model_income)
model_performance(midpoint_concept_model_income)
summary(midpoint_concept_model_income)
summ(midpoint_concept_model_income, digits = 3)
report(midpoint_concept_model_income, digits = 3)

# Only mother education
midpoint_concept_model_mother <- 
  lm(midpoint ~ exposure_balance * 
         correct_mother_education,
     data = concept_df)

check_model(midpoint_concept_model_mother)
model_performance(midpoint_concept_model_mother)
summary(midpoint_concept_model_mother)
summ(midpoint_concept_model_mother, digits = 3)
report(midpoint_concept_model_mother, digits = 3)

# # Only father education
# midpoint_concept_model_father <- 
#   lm(midpoint ~ exposure_balance + 
#          correct_father_education,
#      data = concept_df)
# 
# check_model(midpoint_concept_model_father)
# model_performance(midpoint_concept_model_father)
# summary(midpoint_concept_model_father)
# summ(midpoint_concept_model_father, digits = 3)
# report(midpoint_concept_model_father, digits = 3)

# Only birth order
midpoint_concept_model_birth_order <- 
  lm(midpoint ~ exposure_balance + 
         birth_order_categorical,
     data = concept_df)

check_model(midpoint_concept_model_birth_order)
model_performance(midpoint_concept_model_birth_order)
summary(midpoint_concept_model_birth_order)
summ(midpoint_concept_model_birth_order, digits = 3)
report(midpoint_concept_model_birth_order, digits = 3)
```

```{r}
# Concept vocabulary slope model

# All demographic variables
slope_concept_model <- 
  lm(slope ~ exposure_balance +
       correct_sex +
       income_range_average +
       correct_mother_education +
       # correct_father_education + 
       birth_order_categorical, 
     data = concept_df)

check_model(slope_concept_model)
model_performance(slope_concept_model)
summary(slope_concept_model)
summ(slope_concept_model, digits = 3)
report(slope_concept_model, digits = 3)

# Only sex
slope_concept_model_sex <- 
  lm(slope ~ exposure_balance + 
         correct_sex,
       data = concept_df)

check_model(slope_concept_model_sex)
model_performance(slope_concept_model_sex)
summary(slope_concept_model_sex)
summ(slope_concept_model_sex, digits = 3)
report(slope_concept_model_sex, digits = 3)

# Only income
slope_concept_model_income <- 
  lm(slope ~ exposure_balance + 
         income_range_average,
       data = concept_df)

check_model(slope_concept_model_income)
model_performance(slope_concept_model_income)
summary(slope_concept_model_income)
summ(slope_concept_model_income, digits = 3)
report(slope_concept_model_income, digits = 3)

# Only mother education
slope_concept_model_mother <- 
  lm(slope ~ exposure_balance + 
         correct_mother_education,
     data = concept_df)

check_model(slope_concept_model_mother)
model_performance(slope_concept_model_mother)
summary(slope_concept_model_mother)
summ(slope_concept_model_mother, digits = 3)
report(slope_concept_model_mother, digits = 3)

# # Only father education
# slope_concept_model_father <- 
#   lm(slope ~ exposure_balance + 
#          correct_father_education,
#      data = concept_df)
# 
# check_model(slope_concept_model_father)
# model_performance(slope_concept_model_father)
# summary(slope_concept_model_father)
# summ(slope_concept_model_father, digits = 3)
# report(slope_concept_model_father, digits = 3)

# Only birth order
slope_concept_model_birth_order <- 
  lm(slope ~ exposure_balance * 
         birth_order_categorical,
     data = concept_df)

check_model(slope_concept_model_birth_order)
model_performance(slope_concept_model_birth_order)
summary(slope_concept_model_birth_order)
summ(slope_concept_model_birth_order, digits = 3)
report(slope_concept_model_birth_order, digits = 3)
```

```{r}
# Word vocabulary midpoint model

# All demographic variables
midpoint_word_model <- 
  lm(midpoint ~ exposure_balance +
       correct_sex +
       income_range_average +
       correct_mother_education +
       # correct_father_education + 
       birth_order_categorical,
     data = word_df)

check_model(midpoint_word_model)
model_performance(midpoint_word_model)
summary(midpoint_word_model)
summ(midpoint_word_model, digits = 3)
report(midpoint_word_model, digits = 3)

# Only sex
midpoint_word_model_sex <- 
  lm(midpoint ~ exposure_balance * 
         correct_sex,
       data = word_df)

check_model(midpoint_word_model_sex)
model_performance(midpoint_word_model_sex)
summary(midpoint_word_model_sex)
summ(midpoint_word_model_sex, digits = 3)
report(midpoint_word_model_sex, digits = 3)

# Only income
midpoint_word_model_income <- 
  lm(midpoint ~ exposure_balance + 
         income_range_average,
       data = word_df)

check_model(midpoint_word_model_income)
model_performance(midpoint_word_model_income)
summary(midpoint_word_model_income)
summ(midpoint_word_model_income, digits = 3)
report(midpoint_word_model_income, digits = 3)

# Only mother education
midpoint_word_model_mother <- 
  lm(midpoint ~ exposure_balance * 
         correct_mother_education,
     data = word_df)

check_model(midpoint_word_model_mother)
model_performance(midpoint_word_model_mother)
summary(midpoint_word_model_mother)
summ(midpoint_word_model_mother, digits = 3)
report(midpoint_word_model_mother, digits = 3)

# # Only father education
# midpoint_word_model_father <- 
#   lm(midpoint ~ exposure_balance + 
#          correct_father_education,
#      data = word_df)
# 
# check_model(midpoint_word_model_father)
# model_performance(midpoint_word_model_father)
# summary(midpoint_word_model_father)
# summ(midpoint_word_model_father, digits = 3)
# report(midpoint_word_model_father, digits = 3)

# Only birth order
midpoint_word_model_birth_order <- 
  lm(midpoint ~ exposure_balance + 
         birth_order_categorical,
     data = word_df)

check_model(midpoint_word_model_birth_order)
model_performance(midpoint_word_model_birth_order)
summary(midpoint_word_model_birth_order)
summ(midpoint_word_model_birth_order, digits = 3)
report(midpoint_word_model_birth_order, digits = 3)
```

```{r}
# Word vocabulary slope model

# All demographic variables
slope_word_model <- 
  lm(slope ~ exposure_balance +
       correct_sex +
       income_range_average +
       correct_mother_education +
       # correct_father_education + 
       birth_order_categorical,
     data = word_df)

check_model(slope_word_model)
model_performance(slope_word_model)
summary(slope_word_model)
summ(slope_word_model, digits = 3)
report(slope_word_model, digits = 3)

# Only sex
slope_word_model_sex <- 
  lm(slope ~ exposure_balance + 
         correct_sex,
       data = word_df)

check_model(slope_word_model_sex)
model_performance(slope_word_model_sex)
summary(slope_word_model_sex)
summ(slope_word_model_sex, digits = 3)
report(slope_word_model_sex, digits = 3)

# Only income
slope_word_model_income <- 
  lm(slope ~ exposure_balance + 
         income_range_average,
       data = word_df)

check_model(slope_word_model_income)
model_performance(slope_word_model_income)
summary(slope_word_model_income)
summ(slope_word_model_income, digits = 3)
report(slope_word_model_income, digits = 3)

# Only mother education
slope_word_model_mother <- 
  lm(slope ~ exposure_balance + 
         correct_mother_education,
     data = word_df)

check_model(slope_word_model_mother)
model_performance(slope_word_model_mother)
summary(slope_word_model_mother)
summ(slope_word_model_mother, digits = 3)
report(slope_word_model_mother, digits = 3)

# # Only father education
# slope_word_model_father <- 
#   lm(slope ~ exposure_balance + 
#          correct_father_education,
#      data = word_df)
# 
# check_model(slope_word_model_father)
# model_performance(slope_word_model_father)
# summary(slope_word_model_father)
# summ(slope_word_model_father, digits = 3)
# report(slope_word_model_father, digits = 3)

# Only birth order
slope_word_model_birth_order <- 
  lm(slope ~ exposure_balance * 
         birth_order_categorical,
     data = word_df)

check_model(slope_word_model_birth_order)
model_performance(slope_word_model_birth_order)
summary(slope_word_model_birth_order)
summ(slope_word_model_birth_order, digits = 3)
report(slope_word_model_birth_order, digits = 3)
```
