==> 07_manuscript.Rmd <==
---
title: "07_Manuscript"
author: "Miranda Gómez Díaz"
date: "`r Sys.Date()`"
output: word_document
---

```{r}
#Read needed packages
library(tidyverse)
library(RColorBrewer)
library(rempsyc)
library(gridExtra)
library(ggpubr)

# library(formattable)
# library(sp)
```

```{r}
# Read csv files
Eng_keepers_df <- read_csv("data/02_Eng_keepers.csv")
Fr_keepers_df <- read_csv("data/02_Fr_keepers.csv")
LEQ_01_df <- read_csv("data/01_LEQ_columns.csv")
final_exposure_df <- read_csv("data/05_final_exposure_dataset.csv")
final_vocabulary_scores_df <- read_csv("data/05_final_vocabulary_scores_dataset.csv")
step1_df <- read_csv('data/04_analysis_step_1_dataset.csv')
step2_df <- read_csv('data/05_analysis_step_2_dataset.csv')
coefficients_04_df <- read_csv("data/04_analysis_step_1_coefficients.csv")
vocabulary_scores_df <- read_csv("data/04_analysis_step_1_dataset.csv") 
# These babies had been excluded in one but not the other language (they had missing data or exposure change in only one language)
to_be_excluded <- read_csv("data/02_total_excluded.csv")

midpoint_identification <- read_csv('data/05.1_midpoint_identification.csv')
midpoint_identification_sum <- read_csv('data/05.1_midpoint_identification_summary.csv')
# words_at_spurt_distinct <- read_csv('data/05_words_at_apurt_distinct_dataset.csv')
```

```{r}
# Demographics
# English
Eng_demographics_df <- Eng_keepers_df %>% 
  # Remove rows of uncompleted CDIs
  filter(completed == TRUE) %>% 
  # Keep babies from final sample
  filter(study_id %in% step1_df$study_id) %>% 
  # Keep only demographics columns
  select(study_id, n_months, form_filler, sex, birth_order, mother_education, father_education, annual_income, child_hispanic_latino, child_ethnicity) %>% 
  # Check that each baby has always the same answer
  group_by(study_id) %>% 
  mutate(check_sex = case_when(sex != lead(sex) ~ "check"),
         check_income = case_when(annual_income != lead(annual_income) ~ "check"),
         check_latino = case_when(child_hispanic_latino != lead(child_hispanic_latino) ~ "check"),
         check_ethnicity = case_when(child_ethnicity != lead(child_ethnicity) ~ "check")) %>% 
  ungroup() %>% 
  filter(!(study_id == "S120" & child_hispanic_latino == TRUE)) %>% # Remove row with typo (parents said their child is latino only once)
  # Keep one (the first) row per baby
  group_by(study_id) %>% 
  filter(n_months == min(n_months)) %>% 
  select(-c(check_sex:check_ethnicity)) %>%
  ungroup() #HK note: adding ungroup()

# French
Fr_demographics_df <- Fr_keepers_df %>% 
  # Remove rows of uncompleted CDIs
  filter(completed == TRUE) %>% 
  # Keep babies from final sample
  filter(study_id %in% step1_df$study_id) %>% 
  # Keep only demographics columns
  select(study_id, n_months, form_filler, sex, birth_order, mother_education, father_education, annual_income, child_hispanic_latino, child_ethnicity) %>% 
  # Check that each baby has always the same answer
  group_by(study_id) %>% 
  mutate(check_sex = case_when(sex != lead(sex) ~ "check"),
         check_income = case_when(annual_income != lead(annual_income) ~ "check"),
         check_latino = case_when(child_hispanic_latino != lead(child_hispanic_latino) ~ "check"),
         check_ethnicity = case_when(child_ethnicity != lead(child_ethnicity) ~ "check")) %>% 
  ungroup() %>% 
  filter(!(study_id == "S072" & child_hispanic_latino == TRUE)) %>% # Remove row with typo (parents said their child is latino only once)
  # Keep one (the first) row per baby
  group_by(study_id) %>% 
  filter(n_months == min(n_months)) %>% 
  select(-c(check_sex:check_ethnicity)) %>%
  ungroup() #HK note: adding ungroup()

# There are no discrepancies in the entries for sex or ethnicity. 
# For child_hispanic_latino, if only one answer is different, then the majority answer will be kept. For those who have half-and-half answers: keep first answer.
# For discrepancies in annual_income, the income at recruitment will be kept.

EngFr_demographics_df <-
  full_join(Eng_demographics_df, Fr_demographics_df, by = "study_id", suffix = c("_eng", "_fr")) %>% 
  # Keep correct answers if there were different in Eng and Fr. Replace missing info in one language with the info in the other language; if both are available and the same, keep answer in Eng
  mutate(correct_sex = case_when(is.na(sex_eng) ~ sex_fr, 
                                 is.na(sex_fr) ~ sex_eng,
                                 sex_eng == sex_fr ~ sex_eng),
         correct_birth_order = case_when(is.na(birth_order_eng) ~ birth_order_fr, 
                                 is.na(birth_order_fr) ~ birth_order_eng,
                                 birth_order_eng == birth_order_fr ~ birth_order_eng), # S097 and S098 are twins
         correct_mother_education = case_when(is.na(mother_education_eng) ~ mother_education_fr,
                                              is.na(mother_education_fr) ~ mother_education_eng,
                                              mother_education_eng == mother_education_fr ~ mother_education_eng,
                                              mother_education_eng > mother_education_fr ~ mother_education_eng, # Keep Eng answer if it is highest
                                              mother_education_eng < mother_education_fr ~ mother_education_fr), # Keep Fr answer if it is highest
         correct_father_education = case_when(is.na(father_education_eng) ~ father_education_fr,
                                              is.na(father_education_fr) ~ father_education_eng,
                                              father_education_eng == father_education_fr ~ father_education_eng,
                                              father_education_eng > father_education_fr ~ father_education_eng, # Keep Eng answer if it is highest
                                              father_education_eng < father_education_fr ~ father_education_fr), # Keep Fr answer if it is highest
         correct_ethnicity = case_when(is.na(child_ethnicity_eng) ~ child_ethnicity_fr, 
                                 is.na(child_ethnicity_fr) ~ child_ethnicity_eng, # A French-monolingual baby never reported ethnicity (has NA in both languages)
                                 child_ethnicity_eng == child_ethnicity_fr ~ child_ethnicity_eng),
         correct_income = case_when(#is.na(annual_income_eng) & is.na(annual_income_eng) ~ NA,
                             is.na(annual_income_eng) ~ annual_income_fr,
                             is.na(annual_income_fr) ~ annual_income_eng,
                             annual_income_eng == "Prefer not to disclose" & !is.na(annual_income_fr) ~ annual_income_eng,
                             annual_income_fr == "Prefer not to disclose" & !is.na(annual_income_eng) ~ annual_income_fr,
                             annual_income_eng > annual_income_fr ~ annual_income_eng, # Keep Eng answer if it is highest
                             annual_income_eng < annual_income_fr ~ annual_income_fr, # Keep Fr answer if it is highest
                             annual_income_eng == annual_income_fr ~ annual_income_eng),
         # Convert income ranges into a facor. "Prefer not to disclose" will be considered NA.
         income_factor = case_when(correct_income == "25000-50000" ~ 1,
                                   correct_income == "50000-75000" ~ 2,
                                   correct_income == "75000-100000" ~ 3 ,
                                   correct_income == "100000-125000" ~ 4,
                                   correct_income == "125000-150000" ~ 5,
                                   correct_income == "150000-175000" ~ 6,
                                   correct_income == "175000-200000" ~ 7,
                                   correct_income == ">200000" ~ 8),
         # Estimate parental education
         parental_education = case_when(is.na(correct_father_education) ~ correct_mother_education,
                                        TRUE ~ (correct_mother_education + correct_father_education) / 2)
         ) 

# Demographics: Sex 
demographics_sex_summary <- EngFr_demographics_df %>% 
  ungroup() %>% 
  count(correct_sex) %>% 
  mutate(percent_sex = round((n * 100 / sum(n)), digits = 2))

# Demographics: Sex 
demographics_income_summary <- EngFr_demographics_df %>% 
  ungroup() %>% 
  count(correct_income) %>% 
  mutate(percent_income = round((n * 100 / sum(n)), digits = 2))

# Demographics: parental education
demographics_ses_summary <- EngFr_demographics_df %>% 
  ungroup() %>% 
  summarise(mother_education_M = round(mean(correct_mother_education, na.rm = TRUE), digits = 2),
            mother_education_SD = round(sd(correct_mother_education, na.rm = TRUE), digits = 2),
            father_education_M = round(mean(correct_father_education, na.rm = TRUE), digits = 2),
            father_education_SD = round(sd(correct_father_education, na.rm = TRUE), digits = 2))

# Demographics: Ethnicity 
demographics_ethnicity_summary <- EngFr_demographics_df %>% 
  ungroup() %>% 
  count(correct_ethnicity) %>% 
  mutate(percent_ethnicity = round((n * 100 / sum(n)), digits = 2))

# Save dataframe as csv file. It will be used in Script 05.3 to control for demographic variables
write_csv(EngFr_demographics_df, 'data/07_demographics_dataset.csv')
```

```{r}
# Form filler demographics
# English
Eng_formfiller_df <- Eng_keepers_df %>% 
  # Remove rows of uncompleted CDIs
  filter(completed == TRUE) %>% 
  # Keep babies from final sample
  filter(study_id %in% step1_df$study_id) %>% 
  # Keep only needed columns
  select(study_id, n_months, form_filler)

# French
Fr_formfiller_df <- Fr_keepers_df %>% 
  # Remove rows of uncompleted CDIs
  filter(completed == TRUE) %>% 
  # Keep babies from final sample
  filter(study_id %in% step1_df$study_id) %>% 
  # Keep only needed columns
  select(study_id, n_months, form_filler)

EngFr_formfiller_df <-
  full_join(Eng_formfiller_df, Fr_formfiller_df, by = c("study_id", "n_months"), suffix = c("_eng", "_fr")) %>% 
  mutate(month_filler = case_when(form_filler_eng == form_filler_fr ~ form_filler_eng, # Same parent filled both CDIs
                                  is.na(form_filler_eng) ~ form_filler_fr, # Only French CDI was filled (monolinguals or missing data)
                                  is.na(form_filler_fr) ~ form_filler_eng, # Only English CDI was filled (monolinguals or missing data)
                                  form_filler_eng != form_filler_fr ~ "different caregiver")) %>% 
  group_by(month_filler) %>% 
  summarise(N = n()) %>% 
  ungroup() %>% 
  mutate(percent = round(N * 100 / sum(N), digits = 0))
```

```{r}
# Language status descriptives

# Determine language status of keepers
lang_status_df <- step1_df %>% 
  # Join coefficients and language exposure dataframes
  # Use left_join because LEQ dataset has full sample (N=125), as babies were not excluded from it
  left_join(LEQ_01_df, by = "study_id") %>% 
  group_by(study_id) %>% 
  summarize(lang_status = first(lang_status),
            cumulative_exp_Eng = first(cumulative_exp_Eng),
            cumulative_exp_Fre = first(cumulative_exp_Fre),
            cumulative_exp_L3 = first(cumulative_exp_L3),
            cumulative_exp_L4 = first(cumulative_exp_L4),
            starting_age = first(starting_age),
            lang_status = first(lang_status),
            dominant_lang = first(dominant_lang)) %>% 
  # Set exposure balance as the exposure to the non-dominant language (lower cumulative exposure)
  mutate(exposure_balance = case_when(
    cumulative_exp_Eng < cumulative_exp_Fre ~ cumulative_exp_Eng,
    cumulative_exp_Fre < cumulative_exp_Eng ~ cumulative_exp_Fre))
```

```{r}
# Sample descriptives
sample_descriptives <- step1_df %>% 
# Keep only one row per infant
  group_by(study_id) %>%
  filter(age == last(age),
         vocabulary_type_step1 == last(vocabulary_type_step1)) %>%
  ungroup() %>%
  # Join with LEQ data
  full_join(lang_status_df, by = c("study_id", "starting_age")) %>% 
  summarise(mean_startage = mean(starting_age),
            min_startage = min(starting_age),
            max_startage = max(starting_age),
            sd_startage = sd(starting_age),
            mean_datapts = mean(num_datapts),
            min_datapts = min(num_datapts),
            max_datapts = max(num_datapts),
            sd_datapts = sd(num_datapts),
            mean_missing_datapts = round(mean(num_missing_datapts), digits = 2),
            min_missing_datapts = min(num_missing_datapts),
            max_missing_datapts = max(num_missing_datapts),
            sd_missing_datapts = round(sd(num_missing_datapts), digits =2),
            mean_expbal = mean(exposure_balance),
            min_expbal = min(exposure_balance),
            max_expbal = max(exposure_balance),
            sd_expbal = sd(exposure_balance))

missing_datapts_descriptives <- step1_df %>% 
  # Keep only one row per infant
  group_by(study_id, vocabulary_type_step1) %>%
  filter(age == last(age),
         # vocabulary_type_step1 == last(vocabulary_type_step1)
         ) %>%
  ungroup() %>%
  # # Join with LEQ data
  # full_join(lang_status_df, by = c("study_id", "starting_age")) %>% 
  summarise(mean_missing_datapts = round(mean(num_missing_datapts), digits = 2),
            min_missing_datapts = min(num_missing_datapts),
            max_missing_datapts = max(num_missing_datapts),
            sd_missing_datapts = round(sd(num_missing_datapts), digits =2))
  
total_num_datapts <- step1_df %>% 
  # Keep only one row per infant
  group_by(study_id) %>%
  filter(age == last(age),
         vocabulary_type_step1 == last(vocabulary_type_step1)) %>%
  ungroup() %>% 
  summarise(total_datapts = sum(num_datapts))
```

```{r}
# Descriptives across measures

general_descriptives <- step2_df %>% 
  summarise(midpoint.M = round(mean(midpoint), digits = 2),
            midpoint.SD = round(sd(midpoint), digits = 2),
            slope.M = round(mean(slope), digits = 2),
            slope.SD = round(sd(slope), digits = 2),
            words_learned.M = round(mean(words_learned), digits = 2),
            words_learned.SD = round(sd(words_learned), digits = 2),)
```

```{r}
# TABLE 1. DESCRIPTIVES TABLE

# Midpoint, slope, and words learned at midpoint descriptives

infants_by_vocabtype <- coefficients_04_df %>% 
  # Remove word and concept data from babies who were excluded in one language due to missing data or exposure change in that language
  filter(!((vocabulary_type_step1 == "word" | vocabulary_type_step1 == "concept") & study_id %in% to_be_excluded$study_id)) %>%
  group_by(vocabulary_type_step1) %>% 
  summarise(all_infants = n(),
            )

table_n <- step2_df %>% 
  group_by(vocabulary_type_step1) %>% 
  summarise(N = n(),
            ) %>% 
  full_join(infants_by_vocabtype, by = "vocabulary_type_step1") %>% 
  mutate(percent_infants = round((N*100/all_infants), digits = 0)) #%>% 
  # pivot_wider(names_from = vocabulary_type_step1, values_from = c(N:percent_infants))

table_descriptives_df <- step2_df %>%
  # Remove outlier
  filter(!(study_id == "S125")) %>%
  group_by(study_id) %>%
  select(vocabulary_type_step1, slope, midpoint, words_learned) %>% 
  rename("Age of steepest vocabulary growth (Midpoint)" = midpoint,
         "Steepness of curve (Slope)" = slope,
         "Words learned during month of spurt" = words_learned) %>% 
  pivot_longer(cols = "Steepness of curve (Slope)":"Words learned during month of spurt", names_to = "Measure", values_to = "value") %>% 
  pivot_wider(names_from = vocabulary_type_step1, values_from = value) %>% 
  group_by(Measure) %>% 
    summarise(English.M = round(mean(English, na.rm = TRUE), digits = 2),
            English.SD = round(sd(English, na.rm = TRUE), digits = 2),
            English.Min = round(min(English, na.rm = TRUE), digits = 2),
            English.Max = round(max(English, na.rm = TRUE), digits = 2),
            French.M = round(mean(French, na.rm = TRUE), digits = 2),
            French.SD = round(sd(French, na.rm = TRUE), digits = 2),
            French.Min = round(min(French, na.rm = TRUE), digits = 2),
            French.Max = round(max(French, na.rm = TRUE), digits = 2),
            word.M = round(mean(word, na.rm = TRUE), digits = 2),
            word.SD = round(sd(word, na.rm = TRUE), digits = 2),
            word.Min = round(min(word, na.rm = TRUE), digits = 2),
            word.Max = round(max(word, na.rm = TRUE), digits = 2),
            concept.M = round(mean(concept, na.rm = TRUE), digits = 2),
            concept.SD = round(sd(concept, na.rm = TRUE), digits = 2),
            concept.Min = round(min(concept, na.rm = TRUE), digits = 2),
            concept.Max = round(max(concept, na.rm = TRUE), digits = 2)) %>% 
  mutate("English (n = 19, 51%).M (SD)" = paste0(English.M, " (",English.SD,")"),
         "English (n = 19, 51%).Range" = paste0(English.Min," - ",English.Max),
         "French (n = 26, 72%).M (SD)" = paste0(French.M, " (",French.SD,")"),
         "French (n = 26, 72%).Range" = paste0(French.Min," - ",French.Max),
         "concept (n = 32, 73%).M (SD)" = paste0(concept.M, " (",concept.SD,")"),
         "concept (n = 32, 73%).Range" = paste0(concept.Min," - ",concept.Max),
         "word (n = 32, 73%).M (SD)" = paste0(word.M, " (",word.SD,")"),
         "word (n = 32, 73%).Range" = paste0(word.Min," - ",word.Max)) %>% 
  select(Measure,"English (n = 19, 51%).M (SD)":"word (n = 32, 73%).Range") #HK note: Just a note that after rerunning all scripts/resaving all the data after this code review, the column names in these lines might have to be updated, since they are created manually based off of another table.

table_descriptives <- nice_table(
  table_descriptives_df, 
  separate.header = TRUE, 
  # italics = seq(table_descriptives_df),
  title = c("Table 1", "Descriptive Statistics of the Timing and Magnitude of the Vocabulary Spurt"),
  note = c(
    "Due to sample characteristics and exclusion criteria, each vocabulary type had different group sizes; the percentage next to group size refers to the percentage of infants that spurted in that vocabulary type."
  )
)

table_descriptives
print(table_descriptives, preview ="docx")


```

```{r}
# Model fit table
model_fit_table <- coefficients_04_df %>% 
  select(!c(slope, upper_limit)) %>% 
  pivot_longer(!c(study_id, vocabulary_type_step1, midpoint), names_to = "model", values_to = "AIC") %>% 
  group_by(study_id, vocabulary_type_step1) %>% 
  mutate(best_fit_model = case_when(AIC == min(AIC) ~ model)) %>% 
  filter(!is.na(best_fit_model)) %>% 
  ungroup() %>% 
  group_by(vocabulary_type_step1, best_fit_model) %>% 
  summarise(N = n()) %>% 
  pivot_wider(names_from = best_fit_model, values_from = N)
```

```{r}
exposure_by_vocabtype <- final_exposure_df %>% 
  select(study_id, cumulative_exp_Eng, cumulative_exp_Fre, exposure_balance) %>% 
  rename(English = cumulative_exp_Eng,
         French = cumulative_exp_Fre,
         concept = exposure_balance) %>% 
  mutate(word = concept) %>% 
  pivot_longer(!study_id, names_to = "vocabulary_type_step1", values_to = "exposure")
```

```{r}
# Estimate vocabulary size by age with logistic equation (only spurters)

#Join coefficients and full vocabulary data
logistic_spurters_df <- coefficients_04_df %>%
   # Join vocabulary scores dataframe
  full_join(vocabulary_scores_df, by = c("study_id", "vocabulary_type_step1")) %>%
  group_by(study_id) %>%
  # Remove babies with midpoint out of age range
  filter(midpoint>=16 & midpoint<=max(age)) %>%
  # Remove baby determined as non-spurter
  filter(!(study_id == "S125")) %>%
  # # Remove slope outliers
  # filter(slope < 22) %>%
  # Remove word and concept data from babies who were excluded in one language due to missing data or exposure change in that language
  filter(!((vocabulary_type_step1 == "word" | vocabulary_type_step1 == "concept") & study_id %in% to_be_excluded$study_id)) %>%
  mutate(id_voc = paste0(study_id, vocabulary_type_step1)) %>% # This will be used to filter out data in another df that was removed here
  # Round midpoint to one decimal place
  mutate(midpoint = round(midpoint, digits = 2))

# Create dataframe with age for the logistic curve
ages <- tibble(age = seq(16, 30, by = 0.01))
id_voc_spurters <- logistic_spurters_df %>% distinct(study_id, vocabulary_type_step1, id_voc)
id_voc_ages_spurters <- merge(id_voc_spurters, ages) 

# Create the logistic curve with the equation's parameters
curve_spurters <- id_voc_ages_spurters %>% 
  left_join(coefficients_04_df, by = c("study_id", "vocabulary_type_step1")) %>% 
  mutate(vocabulary_type_step1 = factor(vocabulary_type_step1,
                levels=c("English", "French", "concept", "word"))) %>%
  mutate(vocabulary_score = upper_limit / (1 + exp(-slope*(log(age) - log(midpoint)))),
         centered_age = age - midpoint, 
         midpoint = round(midpoint, digits = 2)) %>% 
  # Join with exposure data
  left_join(exposure_by_vocabtype, by = c("study_id", "vocabulary_type_step1"))
```

```{r}
# Create df with vocabulary score at midpoint (to plot midpoint on individual logistic curves)
vocab_at_midpoint <- curve_spurters %>% 
  filter(age == midpoint)

# Color palette
num_colors <- length(unique(curve_spurters$study_id))
my_palette <- colorRampPalette(brewer.pal(8, "Paired"))(num_colors)
single_palette <- c("steelblue3", "blue4")

# Plot individual logistic curves 
individual_logistic_plot <- curve_spurters %>% 
  ggplot() +
  aes(x = age, y = vocabulary_score, 
      color = study_id,
      group = study_id
      ) +
  geom_line(show.legend = FALSE) +
  geom_point(data = vocab_at_midpoint, aes(x = midpoint, y = vocabulary_score), show.legend = FALSE) +
  labs(x = "Age (Months)", y = "Vocabulary Size (Words)") +
  # scale_color_manual(values = my_palette) +
  theme_bw() +
  facet_wrap(vars(vocabulary_type_step1), ncol = 2, scales = "free") 

individual_logistic_plot

# Single language
singlelang_individual_logistic_plot <- curve_spurters %>% 
  filter(vocabulary_type_step1 == "English" | vocabulary_type_step1 == "French") %>% 
  ggplot() +
  aes(x = age, y = vocabulary_score, 
      color = exposure,
      group = study_id
      ) +
  geom_line(show.legend = TRUE) +
  geom_point(data = vocab_at_midpoint %>% filter(vocabulary_type_step1 == "English" | vocabulary_type_step1 == "French"), aes(x = midpoint, y = vocabulary_score), show.legend = FALSE) +
  labs(x = "Age (Months)", y = "Vocabulary Size (Words)", color = "Exposure percent") +
  scale_color_gradient(low = "#ffc0c0", high = "#312271") +
  theme_bw() +
  theme(legend.position="bottom") +
  facet_wrap(vars(vocabulary_type_step1), nrow = 2, scales = "free") 

singlelang_individual_logistic_plot

# Combined languages
combinedlang_individual_logistic_plot <- curve_spurters %>% 
  filter(vocabulary_type_step1 == "concept" | vocabulary_type_step1 == "word") %>% 
  ggplot() +
  aes(x = age, y = vocabulary_score, 
      color = exposure,
      group = study_id
      ) +
  geom_line(show.legend = TRUE) +
  geom_point(data = vocab_at_midpoint %>% filter(vocabulary_type_step1 == "concept" | vocabulary_type_step1 == "word"), aes(x = midpoint, y = vocabulary_score), show.legend = FALSE) +
  labs(x = "Age (Months)", y = "Vocabulary Size (Words)", color = "Exposure balance") +
  scale_color_gradient(low = "#9bb306", high = "#00366c") +
  theme_bw() +
  theme(legend.position="bottom") +
  facet_wrap(vars(vocabulary_type_step1), nrow = 2, scales = "free") 

combinedlang_individual_logistic_plot

# Combine plots
all_logistic_plots <- ggarrange(singlelang_individual_logistic_plot, combinedlang_individual_logistic_plot + rremove("ylab"), ncol = 2, 
                                # nrow = 2, 
                                legend = "bottom" 
                                # common.legend = TRUE
                                ) 

all_logistic_plots #HK note: the plots look good, but it could be confusing for readers that the Word plot has a different y-axis scale than the others... if they all went up to 1000 or 1500, then it would be clearer with them side by side that Word vocab is the biggest, and the individual language ones are the smallest. Just an idea, though.

ggsave(file="figures/logistic_curve_plots.jpg", all_logistic_plots, width = 16, height = 12, units = "cm")

```

```{r}
# Plot midpoint and slope models for single-language vocabularies
Sing_midpoint_plot <- step2_df %>%
  filter(vocabulary_type_step2 == "single") %>%
  # filter(!(study_id == "S101" & vocabulary_type_step1 == "French")) %>%
  mutate(vocab_type = case_when(vocabulary_type_step2 == "single" ~ "Single-Language")) %>%
  ggplot(aes(x = exposure_percent, y = midpoint)) +
  geom_point(color = "#e31a1c") +
  stat_smooth(method = lm, alpha = 0.15, color = "#e31a1c") +
  labs(x = "Exposure percent", y = "Timing of spurt (midpoint)") +
  facet_grid(~vocab_type) +
  # scale_color_manual(values = WoCo_palette) +
  xlim(10,100) +
  ylim(16,30) +
  theme_bw() +
  theme(legend.position = "bottom")
   

Sing_midpoint_plot

Sing_slope_plot <- step2_df %>%
  filter(vocabulary_type_step2 == "single") %>%
  # filter(!(study_id == "S101" & vocabulary_type_step1 == "French")) %>%
  mutate(vocab_type = case_when(vocabulary_type_step2 == "single" ~ "Single-Language")) %>%
  ggplot(aes(x = exposure_percent, y = slope)) +
  geom_point(color = "#e31a1c") +
  stat_smooth(method = lm, alpha = 0.15, color = "#e31a1c") +
  labs(x = "Exposure percent", y = "Steepness (slope)") +
  facet_grid(~vocab_type) +
  # scale_color_manual(values = WoCo_palette) +
  xlim(10,100) +
  ylim(0,20) +
  theme_bw() +
  theme(legend.position = "bottom")


Sing_slope_plot

# Plot midpoint and slope models for word and concept vocabularies
WoCo_palette <- c("#1f78b4", "#ff7f00")

WoCo_midpoint_plot <- step2_df %>%
  filter(vocabulary_type_step2 == "word" | vocabulary_type_step2 == "concept") %>%
  mutate(vocab_type = case_when(vocabulary_type_step2 == "word" ~ "Word",
                                vocabulary_type_step2 == "concept" ~ "Concept")) %>% 
  mutate(vocab_type = factor(vocab_type,
                levels=c("Concept", "Word"))) %>%
  ggplot(aes(x = exposure_balance, y = midpoint, color = vocab_type)) +
  geom_point(show.legend = FALSE) +
  stat_smooth(method = lm, alpha = 0.15, show.legend = FALSE) +
  scale_color_manual(values = WoCo_palette) +
  labs(x = "Exposure balance", y = "Timing of spurt (midpoint)") +
  facet_grid(~vocab_type) +
  ylim(16,30) +
  theme_bw() +
  theme(legend.position = "none")

WoCo_midpoint_plot

WoCo_slope_plot <- step2_df %>%
  filter(vocabulary_type_step2 == "word" | vocabulary_type_step2 == "concept") %>%
  mutate(vocab_type = case_when(vocabulary_type_step2 == "word" ~ "Word",
                                vocabulary_type_step2 == "concept" ~ "Concept")) %>% 
  mutate(vocab_type = factor(vocab_type,
                levels=c("Concept", "Word"))) %>%
  ggplot(aes(x = exposure_balance, y = slope, color = vocab_type)) +
  geom_point() +
  stat_smooth(method = lm, alpha = 0.15,) +
  scale_color_manual(values = WoCo_palette) +
  labs(x = "Exposure balance", y = "Steepness (slope)", color = "Vocabulary type") +
  facet_grid(~vocab_type) +
  ylim(0,20) +
  theme_bw() +
  theme(legend.position="bottom")


WoCo_slope_plot

# Put both graphs together
WoCo_plots <-  arrangeGrob(WoCo_midpoint_plot, WoCo_slope_plot, nrow=1)
WoCo_plots #HK note: this line doesn't produce a plot for me... just a list object. If it also causes problems for you, you could try using the patchwork r package to combine plots easily
# ggsave(file="WoCo_plots.png", WoCo_plots, width = 17, height = 8, units = "cm")


step2_plots <- ggarrange(Sing_midpoint_plot, 
                         WoCo_midpoint_plot + rremove("ylab") + rremove("y.text") + rremove("y.ticks"), 
                         Sing_slope_plot, 
                         WoCo_slope_plot + rremove("ylab") + rremove("y.text") + rremove("y.ticks"), 
                         ncol = 2, nrow = 2, widths = c(1.2, 2), legend = "none", common.legend = FALSE)
step2_plots
ggsave(file="figures/step2_plots.jpg", step2_plots, width = 16, height = 12, units = "cm")

```

```{r}
# Plot words learned for single-language vocabularies
Sing_wordslearned_plot <- step2_df %>%
  filter(vocabulary_type_step2 == "single") %>%
  # filter(!(study_id == "S101" & vocabulary_type_step1 == "French")) %>%
  mutate(vocab_type = case_when(vocabulary_type_step2 == "single" ~ "Single-Language")) %>%
  ggplot(aes(x = exposure_percent, y = words_learned)) +
  geom_point(color = "#e31a1c") +
  stat_smooth(method = lm, alpha = 0.15, color = "#e31a1c") +
  labs(x = "Exposure percent", y = "Word learned during the spurt") +
  facet_grid(~vocab_type) +
  # scale_color_manual(values = WoCo_palette) +
  xlim(10,100) +
  ylim(0,180) +
  theme_bw() +
  theme(legend.position = "bottom")

Sing_wordslearned_plot

# Plot words learned for multi-language vocabularies

WoCo_wordslearned_plot <- step2_df %>%
  filter(vocabulary_type_step2 == "word" | vocabulary_type_step2 == "concept") %>%
  mutate(vocab_type = case_when(vocabulary_type_step2 == "word" ~ "Word",
                                vocabulary_type_step2 == "concept" ~ "Concept")) %>% 
  mutate(across(vocab_type, factor,
                levels=c("Concept", "Word"))) %>%
  ggplot(aes(x = exposure_balance, y = words_learned, color = vocab_type)) +
  geom_point(show.legend = FALSE) +
  stat_smooth(method = lm, alpha = 0.15, show.legend = FALSE) +
  scale_color_manual(values = WoCo_palette) +
  labs(x = "Exposure balance", y = "Timing of spurt (midpoint)") +
  facet_grid(~vocab_type) +
  # xlim(10,100) +
  ylim(0,180) +
  theme_bw() +
  theme(legend.position = "none")

WoCo_wordslearned_plot

# Put both graphs together
wordslearned_plots <- ggarrange(Sing_wordslearned_plot, 
                         WoCo_wordslearned_plot + rremove("ylab") + rremove("y.text") + rremove("y.ticks"), 
                         # Sing_slope_plot, 
                         # WoCo_slope_plot + rremove("ylab") + rremove("y.text") + rremove("y.ticks"), 
                         ncol = 2, 
                         # nrow = 2, 
                         widths = c(1.2, 2), legend = "none", common.legend = FALSE)
wordslearned_plots
ggsave(file="figures/wordslearned_plots.jpg", wordslearned_plots, width = 16, height = 8, units = "cm")
```

```{r}
#  COMPARISON OF SPURTERS VS NON-SPURTERS

# Estimate vocabulary size by age with logistic equation (spurters and non-spurters)

#Join coefficients and full vocabulary data
logistic_full_df <- coefficients_04_df %>%
   # Join vocabulary scores dataframe
  full_join(vocabulary_scores_df, by = c("study_id", "vocabulary_type_step1")) %>%
  #group_by(study_id) %>% #HK note: commenting this out because the rest of this doesn't depend on a grouped dataset
  # # Remove babies with midpoint out of age range
  # filter(midpoint>=16 & midpoint<=max(age)) %>%
  # # Remove slope outliers
  # filter(slope < 22) %>%
  # Remove word and concept data from babies who were excluded in one language due to missing data or exposure change in that language
  filter(!((vocabulary_type_step1 == "word" | vocabulary_type_step1 == "concept") & study_id %in% to_be_excluded$study_id)) %>%
  mutate(id_voc = paste0(study_id, vocabulary_type_step1)) %>% # This will be used to filter out data in another df that was removed here
  # Round midpoint to one decimal place
  mutate(midpoint = round(midpoint, digits = 2)) 

# Create dataframe with age for the logistic curve
# ages <- tibble(age = seq(16, 30, by = 0.01))
id_voc_full <- logistic_full_df %>% distinct(study_id, vocabulary_type_step1, id_voc)
id_voc_ages_full <- merge(id_voc_full, ages) 

# Create the logistic curve with the equation's parameters
curve_full <- id_voc_ages_full %>% 
  left_join(coefficients_04_df, by = c("study_id", "vocabulary_type_step1")) %>% 
  mutate(vocabulary_type_step1 = factor(vocabulary_type_step1,
                levels=c("English", "French", "concept", "word"))) %>%
  mutate(vocabulary_score = upper_limit / (1 + exp(-slope*(log(age) - log(midpoint)))),
         centered_age = age - midpoint, 
         midpoint = round(midpoint, digits = 2)) %>% 
  # Add midpoint identification column
  left_join(midpoint_identification_sum %>% select(study_id, vocabulary_type_step1, midp_identify), by = c("study_id", "vocabulary_type_step1"))
```

```{r}
# Create df with vocabulary size at 18 months
vocab_at_18m <- 
  # curve_full %>% filter(age == 18.00) %>% # Use this line to use vocabulary size estimated by the logistic equation
  midpoint_identification %>% filter(age >= 18 & age < 19) %>%  # Use this line to use actual vocabulary size (there might be babies with missing data)
  group_by(study_id, vocabulary_type_step1) %>%
  filter(row_number() == 1) %>%
  ungroup()

table_n_18m <- vocab_at_18m %>% 
  group_by(vocabulary_type_step1, midp_identify) %>% 
  summarise(N = n()) %>% 
  pivot_wider(names_from = midp_identify, values_from = N) %>% 
  rename(Spurters.N = identified,
         Non_Spurters.N = not_identified)

# Vocabulary size at 18 months descriptives
vocab_at_18m_summary <- vocab_at_18m %>%
  group_by(study_id) %>%
  # ungroup() %>% 
  select(vocabulary_type_step1, vocabulary_score, midp_identify) %>% 
  pivot_wider(names_from = midp_identify, values_from = vocabulary_score) %>%
  group_by(vocabulary_type_step1) %>% 
  summarise(# Spurters.N = length(unique(identified, na.rm = TRUE)),
            # Spurters.N = length(unique(is.numeric(identified))),
            Spurters.M = round(mean(identified, na.rm = TRUE), digits = 0),
            Spurters.SD = round(sd(identified, na.rm = TRUE), digits = 0),
            Spurters.Min = round(min(identified, na.rm = TRUE), digits = 0),
            Spurters.Max = round(max(identified, na.rm = TRUE), digits = 0),
            # Non_Spurters.N = length(unique(not_identified, na.rm = TRUE)),
            Non_Spurters.M = round(mean(not_identified, na.rm = TRUE), digits = 0),
            Non_Spurters.SD = round(sd(not_identified, na.rm = TRUE), digits = 0),
            Non_Spurters.Min = round(min(not_identified, na.rm = TRUE), digits = 0),
            Non_Spurters.Max = round(max(not_identified, na.rm = TRUE), digits = 0)
            ) %>% 
  full_join(table_n_18m, by = "vocabulary_type_step1") %>% 
    mutate("18 Months.Spurters.M (SD)" = paste0(Spurters.M, " (",Spurters.SD,")"),
         "18 Months.Spurters.Range" = paste0(Spurters.Min," - ",Spurters.Max),
         "18 Months.Non-Spurters.M (SD)" = paste0(Non_Spurters.M, " (",Non_Spurters.SD,")"),
         "18 Months.Non-Spurters.Range" = paste0(Non_Spurters.Min," - ",Non_Spurters.Max)) %>% 
  select(vocabulary_type_step1, Non_Spurters.N, "18 Months.Non-Spurters.M (SD)", "18 Months.Non-Spurters.Range", Spurters.N,"18 Months.Spurters.M (SD)", "18 Months.Spurters.Range") %>% 
  rename("18 Months.Spurters.N" = Spurters.N,
         "18 Months.Non-Spurters.N" = Non_Spurters.N) #HK note: just flagging these manual column names to adjust later with finalized numbers
  

table_18m_descriptives <- nice_table(
  vocab_at_18m_summary, 
  separate.header = TRUE, 
  # italics = seq(table_descriptives_df),
  title = c("Table 2", "Descriptive Statistics of Vocabulary Size at 18 Months"),
  note = c(
    "Descriptive statistics of spurters and non-spurters in each vocabulary type."
  )
)

table_18m_descriptives
# print(table_18m_descriptives, preview ="docx")

```

```{r}
# Create df with vocabulary size at 30 months
vocab_at_30m <- 
  # curve_full %>% filter(age == 30.00) %>% # Use this line to use vocabulary size estimated by the logistic equation
  # Use these lines to use actual vocabulary size (there might be babies with missing data)
  # midpoint_identification %>% filter(age >= 30 & age < 31) # Some babies last submitted data was at ALMOST 30 months; this approach leaves them out
  # The following approach use each baby's last submitted data (age may vary)
  midpoint_identification %>%
  group_by(study_id, vocabulary_type_step1) %>%
  arrange(desc(age)) %>%
  filter(row_number() == 1,
         # age >= 29.5
         ) %>% 
  group_by(study_id, vocabulary_type_step1) %>%
  filter(row_number() == 1) %>%
  ungroup()

table_n_30m <- vocab_at_30m %>% 
  group_by(vocabulary_type_step1, midp_identify) %>% 
  summarise(N = n()) %>% 
  pivot_wider(names_from = midp_identify, values_from = N) %>% 
  rename(Spurters.N = identified,
         Non_Spurters.N = not_identified)

# Vocabulary size at 30 months descriptives
vocab_at_30m_summary <- vocab_at_30m %>%
  group_by(study_id) %>%
  # ungroup() %>% 
  select(vocabulary_type_step1, vocabulary_score, midp_identify) %>% 
  pivot_wider(names_from = midp_identify, values_from = vocabulary_score) %>%
  group_by(vocabulary_type_step1) %>% 
  summarise(# Spurters.N = length(unique(identified, na.rm = TRUE)),
            # Spurters.N = length(unique(is.numeric(identified))),
            Spurters.M = round(mean(identified, na.rm = TRUE), digits = 0),
            Spurters.SD = round(sd(identified, na.rm = TRUE), digits = 0),
            Spurters.Min = round(min(identified, na.rm = TRUE), digits = 0),
            Spurters.Max = round(max(identified, na.rm = TRUE), digits = 0),
            # Non_Spurters.N = length(unique(not_identified, na.rm = TRUE)),
            Non_Spurters.M = round(mean(not_identified, na.rm = TRUE), digits = 0),
            Non_Spurters.SD = round(sd(not_identified, na.rm = TRUE), digits = 0),
            Non_Spurters.Min = round(min(not_identified, na.rm = TRUE), digits = 0),
            Non_Spurters.Max = round(max(not_identified, na.rm = TRUE), digits = 0)
            ) %>% 
  full_join(table_n_30m, by = "vocabulary_type_step1") %>% 
  # relocate(Spurters.N, .before = Spurters.M) %>% 
  # relocate(Non_Spurters.N, .before = Non_Spurters.M) %>% 
  
  mutate("30 Months.Spurters.M (SD)" = paste0(Spurters.M, " (",Spurters.SD,")"),
         "30 Months.Spurters.Range" = paste0(Spurters.Min," - ",Spurters.Max),
         "30 Months.Non-Spurters.M (SD)" = paste0(Non_Spurters.M, " (",Non_Spurters.SD,")"),
         "30 Months.Non-Spurters.Range" = paste0(Non_Spurters.Min," - ",Non_Spurters.Max)) %>% 
  select(vocabulary_type_step1, Non_Spurters.N, "30 Months.Non-Spurters.M (SD)", "30 Months.Non-Spurters.Range", Spurters.N,"30 Months.Spurters.M (SD)", "30 Months.Spurters.Range") %>% 
  rename("30 Months.Spurters.N" = Spurters.N,
         "30 Months.Non-Spurters.N" = Non_Spurters.N) #HK note: just flagging these manual column names to adjust later with finalized numbers
  

table_30m_descriptives <- nice_table(
  vocab_at_30m_summary, 
  separate.header = TRUE, 
  # italics = seq(table_descriptives_df),
  title = c("Table 3", "Descriptive Statistics of Vocabulary Size at 30 Months"),
  note = c(
    "Descriptive statistics of spurters and non-spurters in each vocabulary type."
  )
)

table_30m_descriptives
# print(table_30m_descriptives, preview ="docx")

```
```{r}
# Create a single table with comoparion of spurters vs non-spurters at 18 and 30 months (previously tables 2 and 3)

spurters_vs_nonspurters <- full_join(vocab_at_18m_summary, vocab_at_30m_summary, by = "vocabulary_type_step1"
                                     # , suffix = c(".18m", ".30m")
                                     ) #%>% 
  # rename_with(~ paste0("18m.", .x, recycle0 = TRUE, any_of(c(ends_with(".18m"))))
  

spurters_vs_nonspurters_descriptives <- nice_table(
  spurters_vs_nonspurters, 
  separate.header = TRUE, 
  # italics = seq(table_descriptives_df),
  title = c("Table 2", "Descriptive Statistics of Vocabulary Size at 18 and 30 Months"),
  note = c(
    "Descriptive statistics of spurters and non-spurters in each vocabulary type."
  )
)

spurters_vs_nonspurters_descriptives
print(spurters_vs_nonspurters_descriptives, preview ="docx")
```
